Add python-keystoneclient to the ansible-runtime venv
In order to reduce the packages required to pip install on to the hosts,
we install the extra library into the ansible-runtime venv so that tasks
interacting with keystone using our keystone module can do so by using
localhost delegation. Using this strategy, the library does not need to
be installed on the target host.

Change-Id: I7213252ac7fe59550379db3fe755996c1349a959