Only implement openrc/clouds.yaml on a designated host
The openrc role will no longer be executed on all OpenStack
service containers/hosts. Instead a single host is designated
through the use of the ``openstack_service_setup_host``
variable. The default is ``localhost`` (the deployment host).

While each playbook previously executed the role, it did not
need to as the role is already in each service role's meta
dependencies.

Needed-By: https://review.openstack.org/568142
Change-Id: I7bb90264ee379db5410e7ec16c5bac1ee2c398dc