Add mount options whenever formatting disks
This change provides mount options for all of the file system types
we use. These mount options will ensure that the underlying file systems
react with the best possible performance when building a test cloud. The
main motivation behind this patch is to help resolve bug 1753790 which
is causing sporatic failure when gating on RAX test instances that have
a secondary data disks. We were providing no options when mounting a
secondary disk which results in using the system defaults which may not
be very optimal under general workloads. The options have been added to
the bootstrap role and will be consistently applied whenever a given FS
is used.

> Part of this chagne is to modify the FS type for /var/lib/nova. We
  were using EXT4 and the recommended FS for VM workloads is XFS.

> The mount options have largely been derived from the following posts,
  https://www.phoronix.com/scan.php?page=article&item=linux414-fs-compare&num=1
  https://www.phoronix.com/scan.php?page=article&item=linux-415-fs&num=1

Change-Id: I75e00af687d46551e19178604b60665342cf2cae
Closes-Bug: #1753790
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>