Remove rally_git_* overrides
Presently, rally in the integrated build is installed from PyPI, but without
constraints. This means the version installed may not be what is expected,
and the libraries that it pulls in are unconstrained and may cause failure
as has been seen recently with the new release of the cmd2 library which is
python3 only.

This commit removes the playbooks/defaults/repo_packages/openstack_testing.yml
file as it only contains the rally git overrides. This results in the rally
venv getting installed from PyPI but with upper-constraints applied.

Needed-By: https://review.openstack.org/570905
Change-Id: I79f37deb3615046e6c944008cc73d5cf2da5bb99