Move database creation into role (ironic)
There is no record for why we implement the database creation outside
of the role in the playbook, when we could do it inside the role.

Implementing it inside the role allows us to reduce the quantity of
group_vars duplicated from the role, and allows us to better document
the required variables in the role. The delegation can still be done
as it is done in the playbook too.

In this patch we remove the group_vars which were duplicated from the
role, and remove the DB setup tasks as they are no longer required.

Change-Id: I9ccf73c5797fd5625d8df3c67420cb6724bbe2e2
Depends-On: https://review.openstack.org/571759