Fix loop variable name for nested loop
The HAProxy service management did not work because the loop variable
"item" was used in a nested loop within an include. Because of this the
haproxy Ansible module acted on an inexistant HAProxy backend.
Unfortunately the module does not throw an error in this case.

The issue is fixed by using loop_control/loop_var in the inner loop to
use another loop variable name.

Change-Id: Ia9d23f8861b53eadd7d4eff4a68683ebcda37dc2
(cherry picked from commit 144b8505636f7065519c6c905d0b184367396d82)