Replace rabbitmq references with oslo messaging
This removes, everywhere applicable, the reference
_rabbitmq_, replacing it with the new oslo
counterpart. Shims for compatibility are explicitly
referenced, so that we can remove them later, when
the roles have been adapted appropriately.

It explicitly puts all the oslo messaging details
in one place, as this is easier to track. It is
applied on the all group, as most of the time, the
details are used by 2 or more roles.

This also updated spice-html5 git repository URL
as their code moved out of github. This also adds
some details in job log collection in order to be
able to gather more information about failures.

Depends-On: https://review.openstack.org/#/c/572413/
Depends-On: https://review.openstack.org/#/c/572565/
Depends-On: https://review.openstack.org/#/c/560574/
Change-Id: I1bab622e22d7d820fa2f774b4df60323b1eb7d1e