Update all SHAs for milestone 3
This patch updates all the roles to the latest available stable
SHA's, copies the release notes from the updated roles into the
integrated repo, updates all the OpenStack Service SHA's, and
updates the appropriate python requirements pins.

We also had to make master branch install HEAD of tempest
master: without it, all the sha bumps of keystone would fail in
tempest, as there are discrepencies between latest keystone
and 18.0.0 tempest (the pip installed version).

In order to install tempest from master branch, the workaround in
the gates is:
- Keeping the same behavior as currently when not on master branch
  (installing tempest from pip packages)
- Installing master of tempest without using enabling the
  repo build for everyone, until a new version of tempest
  is released with the patch [1] included.

[1]: https://github.com/openstack/tempest/commit/7d2b636a30057ed8db8cfd4fe2248f509b3570f1

Depends-On: https://review.openstack.org/#/c/573398/
Depends-On: https://review.openstack.org/#/c/583955/
Change-Id: I1edfc43ac571198fcb592ae8e97dcc85d20940af