Use TCP mode for console if SSL is configured
In commit 95b51a83cd41c71838bd4c99033f2cef042e980b the haproxy
configuration for the nova console was changed to use HTTP mode to
support an HTTP health check. While this is desirable if no SSL is used
for the console or SSL is terminated on haproxy, it breaks SSL in case
it's terminated on the nova console container.

This commit introduces a new variable haproxy_nova_console_http_mode
which is True by default and False if SSL certificates are configured
for nova console.

Change-Id: Ie3b7cfa9377d365af6524f11f95d3e557ce0b3cf