Set cache_timeout when creating nspawn containers
Nspawn containers, when first created, do not have any valid cache and
that causes the playbooks to fail on the first run when the package
"ubuntu-cloud-keyring" is installed on the ubuntu platform. This change
sets the cache timeout to 0 when the openstack_hosts role is executed
within a container during the container create process which will ensure
the cache is updated without fail.

Change-Id: I22299785192c691dbb1262ee6c9873a3c7c91315
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>