Do not install tmux
Without this patch, tmux is installed on all the nodes, but
never used.

This is a problem, because some distributions of those three
families (for example SLE) may have decided to not ship tmux
because they ship some other terminal emulator, for example screen.

This patch solves the problem by removing the package that is not
used in any way.

Change-Id: I29f78be4c692fc276e82c3eb71a237614dd7cca0