Allow a bypass of operating system
Without this patch, it is impossible to deploy an OpenStack-Ansible
cloud on an unsupported operating system.

This is a problem, as an experienced developer cannot bypass this check
when working on an experimental OS.

This patch solves the problem by creating a variable to determine if
tests should run, and run the OS test by default.

Change-Id: Icf70f32826cd096ec65309aff47ca5d5c01f57cc