Switch lxc_container_backing_store to default "dir"
It seems that the `machinectl` backing store is prone to errors and it
is causing instability in the gate since it was introduced.  Also, the
default value for it is "dir" which means 99% of deployments will be
using that value.

We should be gating for the most commonly used case, therefore, by
removing it, we're gating with a much similar environment as the one
used by our users.

Conflicts:
1. tests/roles/bootstrap-host/defaults/main.yml
   The variable does not exist. It is hard-coded in the template instead.

Change-Id: Icc099981f2a216d6c4d10f8a23f56cac32cc932f
(cherry picked from commit 8a18a23511e6aa76ac6f73ca4644252aaf99063d)