Set force=true when creating the RPC users
In order to accommodate the changes we've made in the RPC messaging
systems the force flag must be set to `true` when deploying Rocky
in an upgrade situation. Without this flag set to true existing users
will not have their RPC passwords updated and will result in large
scale, obscure, outages across the cloud.

Change-Id: I81ec30bb089908dfa25b248dc4fe7463553457ce
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>