Update HAP check for nova console
The nova console types all have different ports and pathing that needs
to be checked via the loadbalancer when running in production. This
change updates our current check to use the proper pathing based on the
console type being used. Without this change the logs will become full
from 404 errors which is simply a loadbalancer healthcheck.

To support this change, the variable `nova_consoles` was created which
is a hash of consoles and provides an easy way to refernece the proper
port and pathing without complex case-statements.

Change-Id: If3285caea3f9974e030069eef5a5becc654e7a27
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>