Remove special extra repo setup from AIO bootstrap
The AIO bootstrap should not need to use anything from the special
cloud repositories. It should just simply use the base OS and only
things available in the base OS. Having to add the extra repositories
in the AIO bootstrap would mean that deployers would need to add
them themselves for production, and this is not what our docs tell
them.

The openstack-hosts playbook will set these up correctly when the
time is right.

By removing this we test the promise that deployers don't need these
pre-setup, and we also reduce var definition duplication for the RDO
and OBS repositories.

Change-Id: I2bcf1cb9f9526d61ce0ae9fb092c6b5b3baa3bbb