Move MQ vhost/user creation into role (glance)
There is no record for why we implement the MQ vhost/user creation
outside of the role in the playbook, when we could do it inside the
role.

Implementing it inside the role allows us to reduce the quantity of
group_vars duplicated from the role, and allows us to better document
the required variables in the role. The delegation can still be done
as it is done in the playbook too.

In this patch we remove the group_vars which were duplicated from the
role, and remove the MQ setup tasks as they are no longer required.

We also remove the user_secrets which are now totally unused.

Depends-On: https://review.openstack.org/568517
Change-Id: I366d9f7f7ffb0d6912590520a5ea5a718ab0d9af