Allow inventories with no "properties"
Without this patch, inventories with hosts without the field
"properties" cannot be used. Currently, the group variable
resolution will fail when parsing properties.is_metal, as
properties wouldn't be defined.

This is a problem, because it forces deployers with an external
inventory to add a useless "properties" field for them.

This patch solves the problem by simplifying the conditions,
making them valid if the deployer has defined is_metal in its
inventory or not.

Change-Id: Ic09b455016dd3d4f7945900d13d37fdc86f80134