Be consistent with haproxy group name
Without this patch, two different groups are used
for haproxy: "haproxy" and "haproxy_all".

This is a problem if you want to deploy openstack-ansible
with a static inventory, and only define one of the groups:
"haproxy". You would define the group "haproxy" because it
is already the one used in group_vars. In that case, the
group var resolution for galera_all would fail, as
groups['haproxy_all'] does not exist.

This patch solves the problem by ensuring the only
group used for haproxy is named "haproxy" and not "haproxy_all".

Change-Id: Ic43564a8cbcb1aa30ac9b805c04f409dbf885b2e