Ensure container name doesn't need to be defined
Without this patch, container_name needs to be defined in order
to deployer galera.

This is a problem for people with a static inventory and running
galera on metal, as they would not have defined the non standard
variable "container_name" in their inventory.

This patch solves the problem by having a fallback from
container_name to the inventory_hostname, as they would normally
be equal.

Change-Id: I86c369b0b4c7f3c1a521a85cced1786916de9349