Use a static inventory skeleton
This provides a default static inventory, to complement the
current dynamic inventory.

It provides the ability to staticly define groups and their
hierarchy in an easier way than the dynamic inventory.

It can also be used as a basis for deployers to use their
own inventory in /etc/openstack_deploy/inventory.ini.

Change-Id: If9672e3a5794c36b7735ec3d34dc93f5bf79504b