Disable keepalived ping tests by default
At the moment, the keepalived ping tests are something that is
configured by default which is ideal because it's hardcoded with
a public IP address that might result in a failing deployment
if the environment does not for some reason reach it.

We should leave the option to be there however not have all default
deployments hitting this IP as it can introduce behaviour where
if that IP fails, every single OpenStack Ansible deployment would
fail afterwards.

Closes-Bug: #1672453
Change-Id: I5aec4664e67fb2b3e1c0a2fc9782a4ccaa78a39a