Be consistent with memcached group name
Without this patch, two different groups are used for memcached:
"memcached" and "memcached_all".

This is a problem if you want to deploy openstack-ansible with a
static inventory, and only define one of the groups "memcached".
You would define the group "memcached" because it is already the
one used in the group_vars, and it makes more sense as there is
only a single component in "memcached" (unlike nova, cinder, ...).

This patches solves the problem by ensuring the only group used
for memcached is named "memcached" and not "memcached_all".

Change-Id: Ie7ff9a33df8d0eabdee8307a4f8ef5694c72bdd9