Allow the data-disk fstype and directory to be set
As part of https://review.openstack.org/#/c/552047/ the ability to
specify the directory and fstype of the 2nd data disk was configured
(based on the lxc_container_backend). This wasn't backported, and since
the patch was large it would be impractical to backport the whole patch.

This PR backports only the ability to set the filesystem type and path
when configuring the data-disks, which means that using dir or other
backends will put the lxc containers onto the correct partition.

Additionally, set lxc_container_backing_store: "dir" inside the
bootstrap-hosts role. The default in the user_variables sample is dir in
anycase.

Change-Id: I60ec274d4d406aa27d24e8545085f84c6540f801