Fix IP lookup when no container_networks
Without this patch, the set _network_data directive fails when
container_networks is not defined, and the whole address lookup
fails.

This is a problem because, some inventories might not ship
container_networks, and rely on default behavior to fallback
to ansible_host.

This fixes the issue by simplifying the lookup, and relying on
jmespath behavior to fallback to data when its not found,
reducing the amount of jinja written to find the data.

Change-Id: Ib3fce08290f71d1b8682167c722089f52ad7a301