Move get-ansible-role-requirements to scripts
This playbook is the default clone method for roles in both testing
and production. It originally started as a tests-only role fetcher,
so it belonged in tests then, but now that it is used by default
it should reside in scripts.

Also remove an unnecessary reference to the test-inventory.ini,
since the contents of test-inventory.ini are the contents of the
default base ansible inventory if none is provided (meaning
test-inventory.ini is generally redundant.)

Change-Id: I47628f678098b283526e4fc6204520d85046a512