Change the osa wrapper from a heredoc to a file
The OSA warpper has been a heredoc forever however it makes a lot more
sense to have it be a file that we copy. This should help folks
understand where this comes from and what it's doing.

The environment variable "OSA_WRAPPER_BIN" has been added to the bootstrap
script so that deployers could, if they so choose, override the wrapper
script with something else. While I can't imagine folks needing this all
the time, there have been several occasions where it's been necessary to
augment the wrapper tool locally within an environment. Every-time this has
been needed it creates a maintenance burden on operators over the lifetime
of the deployment and it'd be great if we had an official entry point into
this tooling when needed.

Change-Id: Iaac5a8ff070b68a87e4f2b7cebc37749bf513c29
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>