Fix haproxy checks
In the precedent patch [1], I suggested the removal of the test script
used, but that change was not included in the commit.

Therefore, the wiring is currently done with a ping test script,
but the test script has no ip address to ping (incomplete command),
leading to failures.

Instead we should ping on an IP that's reachable on all cases,
which defaults to the default route, or a localhost, unless
the user decides to define its own ip address (like an upstream
router, or a public IP).

[1]: https://review.openstack.org/#/c/580379/

Change-Id: I2f818fd8ed6600ffc0de9557ad56a769ec09881e