Remove useless group var
Since we moved to install tempest from pip packages,
this variable can go away:

The tempest_git_repo group var is applied when:
- The user builds tempest from source instead of from pip packages
- The user wants to use developer mode for tempest.

For the first case, any override in user space of the user
(extra var, group var) would override the group var value,
making it completely useless.

For the second case, if a deployer wants to use developer mode,
this var would override the defaults of the role, which would
probably be an inconvenience for the deployer: the deployer
would then need to override tempest_developer_constraints, or
override the tempest_git_repo itself. That inconvenience can
be removed if the group_var is removed.

Change-Id: Iec8973105686d448b0284cbaeab625a897066664