Use upstream pypi before the repo is built
Trying to reverse proxy upstream pypi has not turned out to
be very stable, or very useful. We've had many, many reports
of stability issues and the additional complexity for offline
and proxy usage is just not worth it.

Given we already have a mechanism in place to handle using
upstream pypi if the repo server is not there yet, disabling
this should just result in that mechanism kicking in and all
will be well again.

Once the repo is built, the reverse proxy to pypiserver will
then be exclusively used and the upstream pypi proxy is not
necessary anyway.

Change-Id: Ic6e4a2d24040655fdd575477ea7438fd6e4d58d9
Needed-By: https://review.openstack.org/584387
(cherry picked from commit 925273f524ecd888f76ac4ad431fcb1a7f742422)