Do not install linux-image-extra by default
In Ubuntu 14.04, the kernel shipped was lightweight, and some
extra features like using vhost_net, required the package
linux-image-extra.

In Ubuntu 16.04 and above, this is not the case anymore,
as many kernel packages ship with extra features.

vhost_net is generally shipped with the kernel package nowadays
and was the only one documented for the addition of the
linux-image-extra into our roles.

Change-Id: I70fa446c2dd00e12702280388299d189e0f3e5c4
Depends-On: https://review.openstack.org/#/c/585376/
Closes-Bug: 1783342
(cherry picked from commit 8913620375908d9ed014725968b22168b515f66b)