Remove all MQ vhost/user and DB create tasks
Now that all the MQ and database creation tasks are in the roles,
and use appropriate defaults, we can remove all the wiring from
group_vars and the tasks.

To cater to the changes in passwords, we also ensure that the
upgrade tooling renames any existing secrets.

The healthcheck-infrastructure.yml playbook is deliberately left
alone due to it being refactored anyway in
https://review.openstack.org/587408

Change-Id: Ie3960e2e2ac9c0aff0bc36f46182be2fc0a038b3