Replace default pip index check with upper constraints check
Now that the repo container is not reverse proxying pypi, the
default index check passes, but anything that tries to use it
prior to the repo build will fail because the repo server has
no wheels built yet.

Instead of checking whether the index is available we switch
to using the upper constraints file check, because that is
only present once the repo is actually built and there are
therefore wheels available on the pypiserver.

Change-Id: Ia9e2456ce17d23250add00462ea32d03788e0d16