Use repo_build_pip_default_index for the pypi fallback
For offline installs, or when using local mirrors, the
variable repo_build_pip_default_index will be set to the
local pypi mirror.

Instead of falling back to pypi, we should fall back to
whatever the repo build is set to use. This gives a more
consistent and user-friendly experience.

Change-Id: Iddac2cc46aa1eb1466ef9e956c01618ef4deac99