Update ceph galaxy cleanup for monolithic repo
In queens we began using the monolithic ceph-ansible repo to source the ceph roles
instead of the broken out galaxy repos[1]. However, in doing this change, we must
also address the situation where upgrades from Pike already have the galaxy style
roles in place in /etc/ansible/roles. Since the ceph-ansible monolithic roles are
later in the roles search path, we continue using the older Pike-installed roles
for ceph, which breaks things with newer ansible versions in Queens.

We should address this by deleting the older Pike roles during the upgrade so
the ceph-ansible monolithic roles will be used going forward.

[1] https://review.openstack.org/#/c/527929/

Closes-Bug: #1784066
Change-Id: I552ebcaf41f101703f52ca9149b90db458190ddc