Ensure package cache is updated for nspawn containers
Due to differences in the way the containers are created, the
package cache must be forcefully updated after creation. If
this is not done then apt thinks that it's up to date and
fails package installs.

We also ensure that the quota is rescanned to ensure that
it is up to date.

The playbook arrangement is also updated to standardise to
the same pattern as the others.

Depends-On: https://review.openstack.org/593033
Change-Id: I44ea902b0785bdeec4565e926105d24fae18500f
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>