[docs] Add documentation for 'install_method' variable
The 'install_method' variable can be used to deploy OpenStack services
either from source or from distribution packages. Since this variable
is meant to be set during the initial deployment and can't be changed
afterwards, it is important to document it in order to let deployers
know about it.

Implements: blueprint openstack-distribution-packages
Change-Id: I59fe66a5cef45487afbfc090aca93cb567427f7a