Only test for repo availability once
The tasks to test for the availability of the repo,
package cache, etc only need to be done on a single
host - and the facts only need to be set by a single
host. From there on, all hosts will have access to
the facts.

This should save a significant amount of time on
large clusters.

Change-Id: I8bae6d8f10b2fb18ffa4cd3c25e11aa348bf4aeb
(cherry picked from commit 8c80ed88483a08803957964ee64e529f5a630ae2)