Temporarily use tempest master for all builds
A fix [1] was introduced in openstack-ansible to use tempest from
the master branch in the AIO, but respecting the installation
from PyPI everywhere else. This allowed users to chose
the appropriate behaviour for their use cases.

This patch [1] also made sure we didn't have anything else to do
when the patch would land on stable branches, as this would
automatically get disabled.

However, as of today, there is still no new version of tempest,
and our stable/rocky branch has been cut. Therefore the fix that
was in place is useless and another approach needs to be
taken.

We will move to use tempest master for master and stable/rocky
branch until the fix that caused this series of patch to
happen is released [2] for tempest on PyPI.

This patch is required for the RC1 freeze on rocky, however
this is blocked due to a master patch in ceph-ansible.
This should unblock the situation temporarily by using
the sha for ceph-ansible that is planned in the RC1 freeze.

[1]: https://github.com/openstack/openstack-ansible/commit/abe0b22c767444408a10c0c210aa3b70f319dc95
[2]: https://github.com/openstack/tempest/commit/7d2b636a30057ed8db8cfd4fe2248f509b3570f1#diff-5c9acbc10dc9d27b47985cd74ab100f6

Change-Id: I42e0cefb28184964a50f66b18d8eba6d2a04eca7
Closes-Bug: 1786095