Temporarily use tempest master for all builds and fix configs
A fix [1] was introduced in openstack-ansible to use tempest from
the master branch in the AIO, but respecting the installation
from PyPI everywhere else. This allowed users to chose
the appropriate behaviour for their use cases.

This patch [1] also made sure we didn't have anything else to do
when the patch would land on stable branches, as this would
automatically get disabled.

However, as of today, there is still no new version of tempest,
and our stable/rocky branch has been cut. Therefore the fix that
was in place is useless and another approach needs to be
taken.

We will move to use tempest master for master and stable/rocky
branch until the fix that caused this series of patch to
happen is released [2] for tempest on PyPI.

A temporary workaround was also added to ensure ceph can still
be built, by freezing the ceph-ansible role sha.

The checksum fill option, was useful a long time ago, but at this point
its just some lovely technical debt that we carry around like an old
dirty security blanket. This change sets us free from the sins of our
past.

We seem to have a lot of jobs that timeout but seem to be very
close to completing, so let's bump the timeout by 20 minutes.

To make up for it, we drop most of our non-voting jobs because
they are taking up resources and always failing.  This should
decrease the amount of resources we're consuming.

Also, instances are failing to launch when jobs hit any nodepool
VMs inside OVH.

[1]: https://github.com/openstack/openstack-ansible/commit/abe0b22c767444408a10c0c210aa3b70f319dc95
[2]: https://github.com/openstack/tempest/commit/7d2b636a30057ed8db8cfd4fe2248f509b3570f1#diff-5c9acbc10dc9d27b47985cd74ab100f6

Change-Id: I42e0cefb28184964a50f66b18d8eba6d2a04eca7
Closes-Bug: 1786095
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>