zuul: Add ceph jobs for distribution installations
Right now, openSUSE and CentOS are missing ceph jobs because they
timeout. We can leverage the distribution installation method to
introduce new ceph jobs which are more likely to pass. However,
CentOS and Ubuntu are not working for this installation method so we
only add a job for openSUSE for the time being.

Change-Id: Ib74467e170c87372d795abf1c1ca4b22951bde3a