Update ansible to 2.5.8
It also seems that Ansible became more strict about the permissions of
the current working directory[1] so we need to either move our ansible.cfg
to a better location, or get rid of it. In the zuul test environment, the
location is world-writable.

Given that this is an ansible.cfg used for tests only, that its content is
(mostly) duplicated in the wrapper's RC file and that implementation of
avoiding the wrapper was done when we had vars plugins (which we no longer
have), I think it's safe to remove it and use the wrapper instead.

Using the wrapper ensures that everything is set in one place, so it's
easier to find. In the AIO bootstrap we then just make sure we override
things that we do not want to interfere with the AIO bootstrap - then unset
them later so that the defaults get applied from then on.

To also reduce the duplication of the bootstrap-aio.sh script in the
gate-check-commit script, we source bootstrap-aio from gate-check-commit.

[1]: https://github.com/ansible/ansible/commit/ff980afefdbe4ceb828bdb1bb2eef03cf616bf63

Change-Id: I6a673c3c4f228d11a9d93766021409a5b1672f63
(cherry picked from commit 24bd4a781fe1ccce532ba032dd28631a954111d7)