Drop non-voting jobs and increase timeout by 20 minutes
We seem to have a lot of jobs that timeout but seem to be very
close to completing, so let's bump the timeout by 20 minutes.

To make up for it, we drop most of our non-voting jobs because
they are taking up resources and always failing.  This should
decrease the amount of resources we're consuming.

Hopefully, we can bring these jobs back.

Change-Id: I276a022b71fee3b62bd60ae6d47bd22ac3152081