Remove checksum checks/fill, they are not needed any longer
We currently run the agents on bare metal so the workaround for
checksums is no longer necessary.

The checksum fill option, was useful a long time ago, but at this point
its just some lovely technical debt that we carry around like an old
dirty security blanket. This change sets us free from the sins of our
past.

This is a combined backport of https://review.openstack.org/593090
and https://review.openstack.org/591882 which were squashed into
https://review.openstack.org/590778.

Change-Id: I9a8cfd33ad43e50584819b84172d7ba84532f695