Use operating system specific IP utilities
Operating systems seem to have decided to put network utilities
in different places, therefore our network bootstrap code was
not working on CentOS because it was not located in the same
place as Ubuntu.

This fixes it which should bring up the interfaces as needed.

Change-Id: I6bf403b4c7d13e6a084ba29fad37b89893321369
(cherry picked from commit 829b7940bb1135c446de9f31fc8334305554e42f)