Remove bonds from AIO network config
On bionic the network bridges refuse to be up, because the underlying
bond interfaces do not come up.

This change removes the bonds which were really only decorative, and plugs
a dedicated dummy interface into each bridge. This simplifies the config
and the bridges should come up as soon as the dummy interface is attached.

Change-Id: I2662df94929c30190cd5d303588fdc9a99370e87