Bind mount glance cache directory
By default glance has ~10GB available to cache images on within it's
container.  Some images exceed that size causing image uploads to fail.
To work around this we can mount the image cache directory on the host.

This has the same danger as bind mounting the image directory (filling
up the host disk), so it's not a new danger, just one to be called out.

Change-Id: I887180b165a802a8d3635c2ad5b2284a123e230d