Add pinned SHA for networking-ovn repo
Currently the role defines the repo, but there is no SHA pin
in the integrated repo, so the repo always builds from the
head of master. This will cause inconsistency with deployments
but, more important, will eventually fail to build due to
inconsistent requirements with the requirements repo SHA pin.

Also, due to their being no group assigned in the role, this
repo and its wheels are built regardless of whether neutron
is, which causes build failures.

Change-Id: Ie27d10026600ae189790c03a8d380dc0b09131d2