Make sure the branch to track is in YAML
There is no reason to rely on comments in the files
used to bump the SHAs of upstream repositories.

Without this, a contributor can easily add a comment
and break the tooling, which relies on regexps and
well formated files to behave.

Instead, as this is a YAML file, we should rely on
a YAML parser in the release tooling, for more
robustness.

This fixes the issues by adding an extra key
(_git_track_branch) per repository, showing what to track.
This could be used in release tooling, and is not hurting
the existing code, as the extra keys are not loaded:
Those files are only consumed by the py_pkgs lookup, which
won't read this extra key.

Change-Id: I8fe08f389160e7be17dd56afde22173778540855
(cherry picked from commit 2a288e1b1e57d2986b25851b51613452c891f3e0)