Allows second pass of haproxy on upgrades
Because of the way the upgrade script marks, it
fails at the second haproxy-install because a completed
file was written for the first pass of haproxy-install.

Rather than rewrite the logic, the script creates a symlink to
haproxy-install.yml to allow the script to run the second haproxy
and runs it because the filename is unique.  It then removes
the symlink once the operation has completed to keep the tree
clean.

Change-Id: Ica78446119a650a34dd445c0ad39e187f9ff92b7