Revert "Add lxc3 compatibility"
This reverts commit 762463f850e0010f0ec1bbc2bd7689886779a51d.

The commit is reverted because it uses a lookup to resolve the version
of lxc. This is executed on the deploy host and not the targets. The
lxc_container_create role now addresses the differences between lxc2/3
configuration keys dynamically [1] and resolves the installed lxc version
on a per-target basis.

[1] https://review.openstack.org/603494

Change-Id: Id3669777ad50f9e90a08212daa8d9e252b01089b
(cherry picked from commit 11e601fa8c5a1dea8d90c52679744e41a2586195)