Move ara to scripts-library
Deployment jobs may wish to install the ARA callback using the same
methods we use in the gate without using the test-oriented script
gate-check-commit.sh. We can solve this by moving the ARA setup to
a scripts-library function, and then call the function inside
gate-check-commit. This will allow downstream users to source
scripts-library and call the setup_ara function to install ARA
as well.

Change-Id: I789fb68be3f372757a4267e1cd7f8cd3b5cf8101