Remove keystone_admin-back load balancer manipulation
For multi-node builds, we previously shut down the keystone_admin-back
for the node being worked on when making changes to the keystone node.

This back-end no longer exists, and the second loop does not use loop
control to ensure that the common tasks 'item' and the playbook 'item'
do not clash, causing the task to enable the back-end to have no effect
and keystone remains in maintenance mode.

This patch removes the back-end from the list, and also removes the
port reservation which is no longer applicable.

Change-Id: I0d7e4804bdc92ffe7a679060a686e684c01fcd1b