Disable nested virt and set SUSE+Bionic jobs to non-voting
The OpenSUSE jobs are always failing due to some RPM issues and
the Bionic job is always timing out.  Let's remove them for now
to be able to land code.

Also, it looks like OpenSUSE and CentOS are getting failures
when running with nested virt on other providers.  Let's just
get rid of it.

Change-Id: Ifae3425b66254aa655df5b6bb95896da1cccc32a