Use the compute kit + horizon for all distros
The AIO is useful for demonstrating OpenStack, and for testing
it. The primary services required for demonstration are the
Compute Kit [1] and Horizon. This also serves us well for test
purposes. Rather than having a mix between the AIO and 'Base Kit'
in the jobs, we converge them all to the 'AIO' set and reduce
the set to only include these services.

For the 'metal' jobs we leave horizon out. At this time, horizon
binds to all listening addresses and conflicts with nginx. This
will need to be fixed in a subsequent patch.

We also switch all Xenial jobs to Bionic, and sort the services
in each scenario alphabetically to make them easier to scan when
comparing them. All the commented jobs in the zuul templates/project
are removed - they're just noise which makes reviewing things
confusing.

The OpenSUSE jobs are all set to non-voting for now due to
issues with the mirror. The bionic/ceph jobs are set to non-voting
for now while an issue with RGW is worked out that causes a segfault
when running tempest against it on bionic.

Change-Id: I67cfaaf1e0bbffbe4f00a5f145f7e410b80c3eb0