Restart cinder-scheduler instead of reloading
During upgrade testing, I consistently notice
that the cinder-scheduler is running at 100% CPU and
appears to be in a hung state after the upgrade.

Restarting the process clears this up so this commit
changes the cinder-scheduler to a restart instead of
a reload.  This appears to have been an issue
previously but I've only seen the scheduler process
spinning at 100%.

https://bugs.launchpad.net/openstack-ansible/+bug/1709346

Change-Id: Ia0bf98413567267865be0503b08d5a03a9698214