Ensure pip/setuptools/wheel are up to date in the utility venv
To ensure consistency, we make sure that the venv used for the
utility container has an up to date and the right versions of
pip/setuptools/wheel.

Without this, on some distributions, we get an obscure error
stating that the wheels from the repo are not compatible with
the python version.

We also update the repo_build role SHA to include the following
patches for a similar reason:
https://review.openstack.org/615213
https://review.openstack.org/615525
https://review.openstack.org/615876

Finally, due to changes in the rabbitmq yum repo GPG keys, we
also need to update the rabbitmq_server repo SHA to include
the update for that:
https://review.openstack.org/616218

Change-Id: Id66ca539984683d7340b50d3f4364c7836251881