Track a stable branch, not master, for ceph-ansible
http://docs.ceph.com/ceph-ansible/master/ states: "The master branch
should be considered experimental and used with caution." So that
really can't be considered a solid basis for building a cloud on. Use
a stable branch instead.

At the time of this patch, the stable-3.2 branch in ceph-ansible is at
the v3.2.0rc1 tag, so not currently considered fully ready for
production deployments either. However, per discussion with odyssey4me
this is OK for the OSA master branch at this stage in the cycle.

Change-Id: Ibd8f64d9889009cf6e80b92254fd580ff559b1be
Related-Bug: 1802195