Move ARA install to end of bootstrap
setup_ara sets up the callback plugins variable
so as it is now, it gets unset during the role
retrieval.  This commit moves it to the end of
bootstrap so that the ANSIBLE_CALLBACK_PLUGINS
is set.

Change-Id: I46ceb2acf377e38b2dc31ffc4a82b7945f8dd4c5
(cherry picked from commit 93831a60915d899e00c0eaa00a1e8b5511de3e88)