Fix ansible-sshd version
The role ansible-sshd has removed existing tags and replaced them
with equivalents in a different format, e.g. 0.4.5 with v0.4.5, [1].
This is causing a failure when attempting to install the role defined
in this repository.

[1] https://github.com/willshersystems/ansible-sshd/issues/98

In order to help get this through the gate, we also disable nested virt
when in OpenStack-CI due to issues at some providers, as was done in
Ifae3425b66254aa655df5b6bb95896da1cccc32a and we ensure that tests in
OpenStack-CI make use of the reverse proxy, as was done in
Id6982945509a174b0aaa74ccbdfa45adc5cfbb7a

Depends-On: https://review.openstack.org/619695
Depends-On: https://review.openstack.org/619691
Change-Id: Id02f3dfbe29cf4ddbe2f3a1f366d2a917b3960b7