Remove unnecessary octavia scenario AIO bootstrap
The role itself handles the certificate generation and the
download of a pre-built amphora image by default, so there
is no need to do it in the AIO bootstrap. Also, several of
the overrides implemented are the same as the defaults, or
break the role execution, so we remove them to allow the
role defaults to be used instead.

Change-Id: If261478ce83b82e16da954e229bf4e7590ddcd4d
(cherry picked from commit e55960fe3096632ab0083424c2fc705daa55242f)